refactor(test): reorganize ASAPClient coverage tests
Merge coverage gaps into test_client.py with scoped circuit-breaker cleanup.
